+++ src/node/index.ts
@@ -1,4 +1,6 @@
+export * from '../shared/ipc'
 export * from './constans'
+export * from './electron'
 export * from './FileTransport'
 export * from './NodeLogger'
 export * from './TerminalColor'